Jump adjustments
- Ground momentum is now properly applied on frame 1 of jump, rather than frame 2
- Full hop vertical velocity is now calculated identically to short hops (no more balloony FHs)
- Attacking out of spin jumps with Sonic now retains momentum
- G&W sideB now retains momentum lol
