+++ fighters/common/src/function_hooks/mod.rs
@@ -31,10 +31,12 @@ pub fn install() {
     momentum_transfer::install();
     //dash_dancing::install();
 
-    // Handles getting rid of the kill zoom
     unsafe {
+        // Handles getting rid of the kill zoom
         const NOP: u32 = 0xD503201Fu32;
         skyline::patching::patch_data(utils::offsets::kill_zoom_regular(), &NOP);
         skyline::patching::patch_data(utils::offsets::kill_zoom_throw(), &NOP);
+        // Changes full hops to calculate vertical velocity identically to short hops
+        skyline::patching::patch_data(0x6d2188, &0x52800015u32);
     }
 }
