+++ fighters/common/src/function_hooks/momentum_transfer.rs
@@ -32,11 +32,10 @@ pub unsafe fn status_attack_air_hook(fighter: &mut L2CFighterCommon, param_1: L2
     let jump_speed_x_max = WorkModule::get_param_float(boma, hash40("run_speed_max"), 0) * ratio;
 
     let mut l2c_agent = L2CAgent::new(fighter.lua_state_agent);
-    let is_speed_backward = KineticModule::get_sum_speed_x(boma, *KINETIC_ENERGY_RESERVE_ATTRIBUTE_MAIN) * PostureModule::lr(boma) < 0.0;
-    let prev_status_check = [*FIGHTER_STATUS_KIND_JUMP, *FIGHTER_STATUS_KIND_JUMP_SQUAT].contains(&StatusModule::prev_status_kind(boma, 0));
-    let mut new_speed = VarModule::get_float(fighter.object(), vars::common::CURRENT_MOMENTUM).clamp(-jump_speed_x_max, jump_speed_x_max);
+    let new_speed = VarModule::get_float(fighter.object(), vars::common::CURRENT_MOMENTUM).clamp(-jump_speed_x_max, jump_speed_x_max);
 
-    if prev_status_check {
+    if StatusModule::prev_status_kind(boma, 0) == *FIGHTER_STATUS_KIND_JUMP
+    || (fighter_kind == *FIGHTER_KIND_SONIC && StatusModule::prev_status_kind(boma, 0) == *FIGHTER_SONIC_STATUS_KIND_SPIN_JUMP) {
         fighter.clear_lua_stack();
         lua_args!(fighter, FIGHTER_KINETIC_ENERGY_ID_CONTROL, new_speed);
         app::sv_kinetic_energy::set_speed(fighter.lua_state_agent);
@@ -54,7 +53,8 @@ pub unsafe fn change_kinetic_momentum_related(boma: &mut smash::app::BattleObjec
     let prev_status_kind = StatusModule::prev_status_kind(boma, 0);
     let situation_kind = StatusModule::situation_kind(boma);
     let fighter_kind = boma.kind();
-    if [*FIGHTER_KIND_CAPTAIN, *FIGHTER_KIND_FALCO, *FIGHTER_KIND_FOX, *FIGHTER_KIND_GAMEWATCH, *FIGHTER_KIND_WOLF].contains(&fighter_kind) && status_kind == *FIGHTER_STATUS_KIND_SPECIAL_N
+    if (([*FIGHTER_KIND_CAPTAIN, *FIGHTER_KIND_FALCO, *FIGHTER_KIND_FOX, *FIGHTER_KIND_GAMEWATCH, *FIGHTER_KIND_WOLF].contains(&fighter_kind) && status_kind == *FIGHTER_STATUS_KIND_SPECIAL_N)
+    || ( fighter_kind == *FIGHTER_KIND_GAMEWATCH && status_kind == *FIGHTER_STATUS_KIND_SPECIAL_S ))
         && situation_kind == *SITUATION_KIND_AIR && [*FIGHTER_STATUS_KIND_JUMP, *FIGHTER_STATUS_KIND_JUMP_SQUAT].contains(&prev_status_kind) {
         return Some(-1);
     }
